+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+
+## [1.2.0] - 09-12-2020
+### Added
+- crack option
+- beautiful print
+- progress bar
+### Changed
+- fix cli
+- fix bugs
+- update readme
+- update readthedocs
+
 ## [1.1.3] - 07-12-2020
-### Aded
+### Added
 - codecov
 - pre-commit
 - rawsec inventored